oprava sort

+++ 3TERM/inf3/2_test_precvicenie_2/Main/Main/Student.cpp
@@ -50,6 +50,6 @@ int Student::getPocetA() {
 	return this->pocetA;
 }
 
-int Student::getPriemer() {
+float Student::getPriemer() {
 	return this->priemer;
 }
