+++ 3TERM/inf3/2_test_precvicenie_2/Main/Main/Student.h
@@ -21,6 +21,6 @@ public:
 	void vypis();
 	string zapis();
 	int getPocetA();
-	int getPriemer();
+	float getPriemer();
 };
